easy with that healthcheck

+++ docker_scripts/first_run.sh
@@ -1,19 +1,19 @@
 #!/usr/bin/env bash
 
-set -x
+#set -x
 set -e
 
 cd "$(dirname "$(readlink -f -- "$0")")"
 
 FLAG=$(./flag.sh)
 
-[ -f $FLAG ] &&
-cat $FLAG && { echo -e "not initing again\n"; exit 0; }
+[ -f $FLAG ] && { echo -e "not initing again, $FLAG found.\n"; exit 0; }
 
 rm -rf venv
 ./first_run0.sh
 ./first_run1.sh
 ./first_run1b.sh
 ./first_run2.sh
-
 touch $FLAG
+./first_run999.sh
+
